初步实现电磁功能

+++ speed/Project/21-TwoVAL/Driver/control.h
@@ -4,15 +4,15 @@ typedef struct PID{float P,pout,I,iout,D,dout,OUT;}PID;
 extern  PID PID_SPEED,PID_TURN;
 
 //µÁ¥≈œ‡πÿ
-extern int   AD_val_1_min;
-extern int   AD_val_2_min;
-extern int   AD_val_3_min;
-extern int   AD_val_1_max;
-extern int   AD_val_2_max;
-extern int   AD_val_3_max; 
-extern int AD_val_1,AD_val_2,AD_val_3,AD_val_4;
-extern int dis_AD_val_1,dis_AD_val_2,dis_AD_val_3,dis_AD_val_4;
-extern int disgy_AD_val_1,disgy_AD_val_2,disgy_AD_val_3,disgy_AD_val_4;
+extern float   AD_val_1_min;
+extern float   AD_val_2_min;
+extern float   AD_val_3_min;
+extern float   AD_val_1_max;
+extern float   AD_val_2_max;
+extern float   AD_val_3_max; 
+extern float AD_val_1,AD_val_2,AD_val_3,AD_val_4;
+extern float dis_AD_val_1,dis_AD_val_2,dis_AD_val_3,dis_AD_val_4;
+extern float disgy_AD_val_1,disgy_AD_val_2,disgy_AD_val_3,disgy_AD_val_4;
 
 extern uint8 Style;
 extern int Speed_Filter_Times;
